Rename ActionDispatch::IntegrationTest to TestCase
Move more tests to describing what they test instead of how.
