+++ actionpack/lib/action_controller/test_case.rb
@@ -250,7 +250,7 @@ module ActionController
 
   # Superclass for ActionController functional tests. Functional tests allow you to
   # test a single controller action per test method. This should not be confused with
-  # integration tests (see ActionDispatch::IntegrationTest), which are more like
+  # integration tests (see ActionDispatch::TestCase), which are more like
   # "stories" that can involve multiple controllers and multiple actions (i.e. multiple
   # different HTTP requests).
   #
