+++ actionpack/lib/action_dispatch/http/response.rb
@@ -25,7 +25,7 @@ module ActionDispatch # :nodoc:
   # For example, the following demo integration test prints the body of the
   # controller response to the console:
   #
-  #  class DemoControllerTest < ActionDispatch::IntegrationTest
+  #  class DemoControllerTest < ActionDispatch::TestCase
   #    def test_print_root_path_to_console
   #      get('/')
   #      puts response.body
