+++ actionpack/lib/action_dispatch/testing/performance_test.rb
@@ -4,7 +4,7 @@ module ActionDispatch
   # An integration test that runs a code profiler on your test methods.
   # Profiling output for combinations of each test method, measurement, and
   # output format are written to your tmp/performance directory.
-  class PerformanceTest < ActionDispatch::IntegrationTest
+  class PerformanceTest < ActionDispatch::TestCase
     include ActiveSupport::Testing::Performance
   end
 end
