+++ actionpack/test/controller/rescue_test.rb
@@ -292,7 +292,7 @@ class RescueControllerTest < ActionController::TestCase
   end
 end
 
-class RescueTest < ActionDispatch::IntegrationTest
+class RescueTest < ActionDispatch::TestCase
   class TestController < ActionController::Base
     class RecordInvalid < StandardError
       def message
