+++ actionpack/test/dispatch/request/url_encoded_params_parsing_test.rb
@@ -1,6 +1,6 @@
 require 'abstract_unit'
 
-class UrlEncodedParamsParsingTest < ActionDispatch::IntegrationTest
+class UrlEncodedParamsParsingTest < ActionDispatch::TestCase
   class TestController < ActionController::Base
     class << self
       attr_accessor :last_request_parameters, :last_request_type
