+++ actionpack/test/dispatch/request_id_test.rb
@@ -25,7 +25,7 @@ class RequestIdTest < ActiveSupport::TestCase
   end
 end
 
-class RequestIdResponseTest < ActionDispatch::IntegrationTest
+class RequestIdResponseTest < ActionDispatch::TestCase
   class TestController < ActionController::Base
     def index
       head :ok
