+++ actionpack/test/dispatch/spec_type_test.rb
@@ -2,11 +2,11 @@ require "abstract_unit"
 
 class SpecTypeTest < ActiveSupport::TestCase
   def assert_dispatch actual
-    assert_equal ActionDispatch::IntegrationTest, actual
+    assert_equal ActionDispatch::TestCase, actual
   end
 
   def refute_dispatch actual
-    refute_equal ActionDispatch::IntegrationTest, actual
+    refute_equal ActionDispatch::TestCase, actual
   end
 
   def test_spec_type_resolves_for_matching_acceptance_strings
