+++ railties/lib/rails/generators/rails/plugin_new/templates/test/integration/navigation_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class NavigationTest < ActionDispatch::IntegrationTest
+class NavigationTest < ActionDispatch::TestCase
 <% unless options[:skip_active_record] -%>
   fixtures :all
 <% end -%>
