+++ railties/lib/rails/generators/test_unit/integration/templates/integration_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class <%= class_name %>Test < ActionDispatch::IntegrationTest
+class <%= class_name %>Test < ActionDispatch::TestCase
   # test "the truth" do
   #   assert true
   # end
