+++ railties/test/application/test_test.rb
@@ -40,7 +40,7 @@ module ApplicationTests
       app_file 'test/integration/posts_test.rb', <<-RUBY
         require 'test_helper'
 
-        class PostsTest < ActionDispatch::IntegrationTest
+        class PostsTest < ActionDispatch::TestCase
           def test_index
             get '/posts'
             assert_response :success
