+++ railties/test/generators/integration_test_generator_test.rb
@@ -7,6 +7,6 @@ class IntegrationTestGeneratorTest < Rails::Generators::TestCase
 
   def test_integration_test_skeleton_is_created
     run_generator
-    assert_file "test/integration/integration_test.rb", /class IntegrationTest < ActionDispatch::IntegrationTest/
+    assert_file "test/integration/integration_test.rb", /class IntegrationTest < ActionDispatch::TestCase/
   end
 end
